-extracted pointer styles
