+++ packages/react/src/themes/teams/components/Popup/popupContentStyles.ts
@@ -1,65 +1,14 @@
 import { ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
 import { PopupContentProps } from '../../../../components/Popup/PopupContent'
 import { PopupContentVariables } from './popupContentVariables'
-
-const rtlMapping = {
-  left: 'right',
-  right: 'left',
-}
-
-const getPointerStyles = (
-  v: PopupContentVariables,
-  rtl: boolean,
-  popperPlacement?: PopupContentProps['placement'],
-) => {
-  const placementValue = (popperPlacement || '').split('-', 1).pop()
-  const placement = (rtl && rtlMapping[placementValue]) || placementValue
-
-  const rootStyles = {
-    top: {
-      marginBottom: v.pointerMargin,
-    },
-    right: {
-      marginLeft: v.pointerMargin,
-    },
-    bottom: {
-      marginTop: v.pointerMargin,
-    },
-    left: {
-      marginRight: v.pointerMargin,
-    },
-  }
-  const pointerStyles = {
-    top: {
-      bottom: `-${v.pointerOffset}`,
-      transform: 'rotate(45deg)',
-    },
-    right: {
-      left: `-${v.pointerOffset}`,
-      transform: 'rotate(135deg)',
-    },
-    bottom: {
-      top: `-${v.pointerOffset}`,
-      transform: 'rotate(-135deg)',
-    },
-    left: {
-      right: `-${v.pointerOffset}`,
-      transform: 'rotate(-45deg)',
-    },
-  }
-
-  return {
-    root: rootStyles[placement],
-    pointer: pointerStyles[placement],
-  }
-}
+import getPointerStyles from '../../getPointerStyles'
 
 const popupContentStyles: ComponentSlotStylesInput<PopupContentProps, PopupContentVariables> = {
   root: ({ props: p, variables: v, rtl }): ICSSInJSStyle => ({
     borderRadius: v.borderRadius,
     display: 'block',
 
-    ...(p.pointing && getPointerStyles(v, rtl, p.placement).root),
+    ...(p.pointing && getPointerStyles(v.pointerOffset, v.pointerMargin, rtl, p.placement).root),
   }),
   pointer: ({ props: p, variables: v, rtl }): ICSSInJSStyle => ({
     display: 'block',
@@ -72,7 +21,7 @@ const popupContentStyles: ComponentSlotStylesInput<PopupContentProps, PopupConte
     height: v.pointerSize,
     width: v.pointerSize,
 
-    ...getPointerStyles(v, rtl, p.placement).pointer,
+    ...getPointerStyles(v.pointerOffset, v.pointerMargin, rtl, p.placement).pointer,
   }),
   content: ({ props: p, variables: v }): ICSSInJSStyle => ({
     display: 'block',
